FIXed the button action bindings
Copying over the button layout had some issues with naming, so some of
the names didn't match up with the proper Form code funciton. But I went
through the buttons and re-bound the proper buttons to their actions.
All in all, only 2 of the buttons had incorrect names. My fault, mostly
for not properly naming the buttons "up", "down", etc...